07-flex-layout-2 checkpoint
My own solution. The cards are arranged to single column when the screen is narrow
and have more columns when the screed is wider. The page need
to be scrolled down to see the footer in narrow window, while in wider
window the footer not taking full height of the screen.